+++ tests/spec/test_dagspec.py
@@ -1708,3 +1708,25 @@ tasks:
     expected = ("Expected 'tasks' to contain a list, but got: "
                 "{'not_a_list': 'value'} (an object of type 'dict')")
     assert str(excinfo.value) == expected
+
+
+# NOTE: not sure why this doesn't work on windows
+@pytest.mark.xfail(sys.platform == 'win32',
+                   reason='This doesnt work on windows')
+def test_prioritize_local_modules(tmp_directory, tmp_imports):
+    # if math has been loaded already, remove it
+    if 'math' in sys.modules:
+        del sys.modules['math']
+
+    Path('math.py').write_text("""
+def function(product):
+    pass
+""")
+
+    Path('pipeline.yaml').write_text("""
+tasks:
+    - source: math.function
+      product: output.csv
+""")
+
+    assert DAGSpec('pipeline.yaml')